Fix split shared build
