+++ dbms/src/TableFunctions/CMakeLists.txt
@@ -5,7 +5,7 @@ list(REMOVE_ITEM clickhouse_table_functions_sources ITableFunction.cpp TableFunc
 list(REMOVE_ITEM clickhouse_table_functions_headers ITableFunction.h TableFunctionFactory.h)
 
 add_library(clickhouse_table_functions ${clickhouse_table_functions_sources})
-target_link_libraries(clickhouse_table_functions dbms ${Poco_Foundation_LIBRARY})
+target_link_libraries(clickhouse_table_functions clickhouse_storages_system dbms ${Poco_Foundation_LIBRARY})
 
 if (Poco_SQLODBC_FOUND)
     target_link_libraries (clickhouse_table_functions ${Poco_SQLODBC_LIBRARY})
